third-with-testing/requirements.txt
msal
requests
tzdata
Flask
pytz
python-dateutil